+++ plugin-dev/Source/Sentry/Private/SentrySettings.cpp
@@ -2,10 +2,30 @@
 
 #include "SentrySettings.h"
 
-USentrySettings::USentrySettings(const FObjectInitializer &ObjectInitializer)
-    : Super(ObjectInitializer), InitAutomatically(false),
-      UploadSymbolsAutomatically(false) {
-  DsnUrl = TEXT("");
-  Release = TEXT("");
-  PropertiesFilePath = TEXT("Config/sentry.properties");
+#include "Misc/Paths.h"
+
+USentrySettings::USentrySettings(const FObjectInitializer& ObjectInitializer)
+	: Super(ObjectInitializer)
+	, InitAutomatically(false)
+	, UploadSymbolsAutomatically(false)
+{
+	DsnUrl = TEXT("");
+	Release = TEXT("");
+
+	LoadDebugSymbolsProperties();
+}
+
+void USentrySettings::LoadDebugSymbolsProperties()
+{
+	const FString PropertiesFilePath = FPaths::Combine(FPaths::ProjectDir(), TEXT("sentry.properties"));
+
+	if (FPaths::FileExists(PropertiesFilePath))
+	{
+		FConfigFile PropertiesFile;
+		PropertiesFile.Read(PropertiesFilePath);
+
+		PropertiesFile.GetString(TEXT("Sentry"), TEXT("defaults.project"), ProjectName);
+		PropertiesFile.GetString(TEXT("Sentry"), TEXT("defaults.org"), OrgName);
+		PropertiesFile.GetString(TEXT("Sentry"), TEXT("auth.token"), AuthToken);
+	}
 }