+++ plugin-dev/Source/SentryEditor/SentryEditor.Build.cs
@@ -1,0 +1,30 @@
+﻿// Copyright (c) 2022 Sentry. All Rights Reserved.
+
+using UnrealBuildTool;
+
+public class SentryEditor : ModuleRules
+{
+    public SentryEditor(ReadOnlyTargetRules Target) : base(Target)
+    {
+        PCHUsage = ModuleRules.PCHUsageMode.UseExplicitOrSharedPCHs;
+
+        PublicDependencyModuleNames.AddRange(
+            new string[]
+            {
+                "Core",
+            }
+        );
+
+        PrivateDependencyModuleNames.AddRange(
+            new string[]
+            {
+                "CoreUObject",
+                "Engine",
+                "Slate",
+                "SlateCore",
+                "EditorStyle",
+                "Sentry"
+            }
+        );
+    }
+}